refactoring of New from Computer table
