+++ Program.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Data.Sqlite; //importar o sqlite
 using LabManager.Database;
 using LabManager.Repositories;
+using LabManager.Models;
 
 var databaseConfig = new DatabaseConfig(); // cria objeto de string de conexão
 
@@ -36,19 +37,8 @@ if(modelName == "Computer")
         int id = Convert.ToInt32(args[2]);
         string ram = args[3];
         string processor = args[4];
-
-        var connection = new SqliteConnection("Data Source=database.db");
-        connection.Open(); //ABRIR O ARQUIVO/conexão database.db
-
-
-        var command = connection.CreateCommand(); //comando criado no banco aberto
-        command.CommandText = "INSERT INTO Computers VALUES ($id, $ram, $processor)"; //@ - STRING COM QUEBRA DE LINHA
-        command.Parameters.AddWithValue("$id", id);
-        command.Parameters.AddWithValue("$ram", ram);
-        command.Parameters.AddWithValue("$processor", processor);
-
-        command.ExecuteNonQuery(); //create não devolve nada, se fosse select teria retorno
-        connection.Close(); // fecha a conexão
+        var computer = new Computer(id, ram, processor);
+        computerRepository.Save(computer);
     }
