+++ Repositories/ComputerRepository.cs
@@ -36,4 +36,20 @@ class ComputerRepository //isolar funcionalidade de acesso a dados
 
         return computers;
     }
+
+    public void Save(Computer computer) //tipo que voce criou
+    {
+        var connection = new SqliteConnection(_databaseConfig.ConnectionString);
+        connection.Open(); //ABRIR O ARQUIVO/conexão database.db
+
+
+        var command = connection.CreateCommand(); //comando criado no banco aberto
+        command.CommandText = "INSERT INTO Computers VALUES ($id, $ram, $processor)"; //@ - STRING COM QUEBRA DE LINHA
+        command.Parameters.AddWithValue("$id", computer.Id);
+        command.Parameters.AddWithValue("$ram", computer.Ram);
+        command.Parameters.AddWithValue("$processor", computer.Processor);
+
+        command.ExecuteNonQuery(); //create não devolve nada, se fosse select teria retorno
+        connection.Close(); // fecha a conexão
+    }
 }
